backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Settings / env
pydantic-settings==2.10.1

# Auth / security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# Database
SQLAlchemy==2.0.43
alembic==1.16.5
asyncpg==0.30.0
psycopg[binary]==3.2.10

